+++ From Nothing/Assembly-CSharp.csproj
@@ -63,12 +63,13 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Assets\Scripts\Enemy_01_movement.cs" />
     <Compile Include="Assets\Scripts\FadeAwayObject.cs" />
     <Compile Include="Assets\Scripts\FMODMixer.cs" />
     <Compile Include="Assets\Scripts\Interactable.cs" />
-    <Compile Include="Assets\Scripts\Level_01\Enemy_01_movement.cs" />
-    <Compile Include="Assets\Scripts\Level_01\Level_01_Interact_Manager.cs" />
+    <Compile Include="Assets\Scripts\Level_01_Interact_Manager.cs" />
     <Compile Include="Assets\Scripts\Level01Dialogue.cs" />
+    <Compile Include="Assets\Scripts\Level02Dialogue.cs" />
     <Compile Include="Assets\Scripts\LevelStart2.cs" />
     <Compile Include="Assets\Scripts\LoadScene.cs" />
     <Compile Include="Assets\Scripts\Manager.cs" />
@@ -78,37 +79,37 @@
     <Compile Include="Assets\Scripts\SceneSwitch.cs" />
     <Compile Include="Assets\Scripts\TextControl.cs" />
     <Reference Include="Cinemachine">
-      <HintPath>C:/Users/CodyLarkin/Desktop/From Nothing/Library/ScriptAssemblies/Cinemachine.dll</HintPath>
+      <HintPath>C:/Users/CodyLarkin/Documents/Github_Reed/From_Nothing/From Nothing/Library/ScriptAssemblies/Cinemachine.dll</HintPath>
     </Reference>
     <Reference Include="Unity.ProBuilder">
-      <HintPath>C:/Users/CodyLarkin/Desktop/From Nothing/Library/ScriptAssemblies/Unity.ProBuilder.dll</HintPath>
+      <HintPath>C:/Users/CodyLarkin/Documents/Github_Reed/From_Nothing/From Nothing/Library/ScriptAssemblies/Unity.ProBuilder.dll</HintPath>
     </Reference>
     <Reference Include="Unity.ProBuilder.Editor">
-      <HintPath>C:/Users/CodyLarkin/Desktop/From Nothing/Library/ScriptAssemblies/Unity.ProBuilder.Editor.dll</HintPath>
+      <HintPath>C:/Users/CodyLarkin/Documents/Github_Reed/From_Nothing/From Nothing/Library/ScriptAssemblies/Unity.ProBuilder.Editor.dll</HintPath>
     </Reference>
     <Reference Include="Unity.ProBuilder.AssetIdRemapUtility">
-      <HintPath>C:/Users/CodyLarkin/Desktop/From Nothing/Library/ScriptAssemblies/Unity.ProBuilder.AssetIdRemapUtility.dll</HintPath>
+      <HintPath>C:/Users/CodyLarkin/Documents/Github_Reed/From_Nothing/From Nothing/Library/ScriptAssemblies/Unity.ProBuilder.AssetIdRemapUtility.dll</HintPath>
     </Reference>
     <Reference Include="Unity.TextMeshPro.Editor">
-      <HintPath>C:/Users/CodyLarkin/Desktop/From Nothing/Library/ScriptAssemblies/Unity.TextMeshPro.Editor.dll</HintPath>
+      <HintPath>C:/Users/CodyLarkin/Documents/Github_Reed/From_Nothing/From Nothing/Library/ScriptAssemblies/Unity.TextMeshPro.Editor.dll</HintPath>
     </Reference>
     <Reference Include="Unity.PackageManagerUI.Editor">
-      <HintPath>C:/Users/CodyLarkin/Desktop/From Nothing/Library/ScriptAssemblies/Unity.PackageManagerUI.Editor.dll</HintPath>
+      <HintPath>C:/Users/CodyLarkin/Documents/Github_Reed/From_Nothing/From Nothing/Library/ScriptAssemblies/Unity.PackageManagerUI.Editor.dll</HintPath>
     </Reference>
     <Reference Include="Unity.CollabProxy.Editor">
-      <HintPath>C:/Users/CodyLarkin/Desktop/From Nothing/Library/ScriptAssemblies/Unity.CollabProxy.Editor.dll</HintPath>
+      <HintPath>C:/Users/CodyLarkin/Documents/Github_Reed/From_Nothing/From Nothing/Library/ScriptAssemblies/Unity.CollabProxy.Editor.dll</HintPath>
     </Reference>
     <Reference Include="com.unity.cinemachine.editor">
-      <HintPath>C:/Users/CodyLarkin/Desktop/From Nothing/Library/ScriptAssemblies/com.unity.cinemachine.editor.dll</HintPath>
+      <HintPath>C:/Users/CodyLarkin/Documents/Github_Reed/From_Nothing/From Nothing/Library/ScriptAssemblies/com.unity.cinemachine.editor.dll</HintPath>
     </Reference>
     <Reference Include="Unity.ProBuilder.AddOns.Editor">
-      <HintPath>C:/Users/CodyLarkin/Desktop/From Nothing/Library/ScriptAssemblies/Unity.ProBuilder.AddOns.Editor.dll</HintPath>
+      <HintPath>C:/Users/CodyLarkin/Documents/Github_Reed/From_Nothing/From Nothing/Library/ScriptAssemblies/Unity.ProBuilder.AddOns.Editor.dll</HintPath>
     </Reference>
     <Reference Include="Unity.TextMeshPro">
-      <HintPath>C:/Users/CodyLarkin/Desktop/From Nothing/Library/ScriptAssemblies/Unity.TextMeshPro.dll</HintPath>
+      <HintPath>C:/Users/CodyLarkin/Documents/Github_Reed/From_Nothing/From Nothing/Library/ScriptAssemblies/Unity.TextMeshPro.dll</HintPath>
     </Reference>
     <Reference Include="Unity.Analytics.DataPrivacy">
-      <HintPath>C:/Users/CodyLarkin/Desktop/From Nothing/Library/ScriptAssemblies/Unity.Analytics.DataPrivacy.dll</HintPath>
+      <HintPath>C:/Users/CodyLarkin/Documents/Github_Reed/From_Nothing/From Nothing/Library/ScriptAssemblies/Unity.Analytics.DataPrivacy.dll</HintPath>
     </Reference>
     <Reference Include="UnityEngine.AIModule">
       <HintPath>C:/Program Files/Unity/Editor/Data/Managed/UnityEngine/UnityEngine.AIModule.dll</HintPath>
@@ -312,7 +313,7 @@
       <HintPath>C:/Program Files/Unity/Editor/Data/UnityExtensions/Unity/UnitySpatialTracking/RuntimeEditor/UnityEngine.SpatialTracking.dll</HintPath>
     </Reference>
     <Reference Include="Unity.Analytics.StandardEvents">
-      <HintPath>C:/Users/CodyLarkin/Desktop/From Nothing/Library/PackageCache/com.unity.analytics@3.2.3/AnalyticsStandardEvents/Unity.Analytics.StandardEvents.dll</HintPath>
+      <HintPath>C:/Users/CodyLarkin/Documents/Github_Reed/From_Nothing/From Nothing/Library/PackageCache/com.unity.analytics@3.2.3/AnalyticsStandardEvents/Unity.Analytics.StandardEvents.dll</HintPath>
     </Reference>
     <Reference Include="netstandard">
       <HintPath>C:/Program Files/Unity/Editor/Data/NetStandard/ref/2.0.0/netstandard.dll</HintPath>
