+++ From Nothing/Assets/Animations/lvl2_platform01.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 53228d30053a801458de9c8ad877bf89
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
