+++ From Nothing/Assets/Animations/platformFloating (2).controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eec4b715cd776f241bd4285e756f5047
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
