+++ From Nothing/Assets/Models/clear.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2b44a293f9d605848aa1942d24e36f30
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
