+++ From Nothing/Assets/Prefabs/boots.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0b0aee8986e00114e8aa3b52eacae9ac
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
