+++ From Nothing/Assets/Prefabs/platformFloating (2).prefab
@@ -244,8 +244,8 @@ GameObject:
   - component: {fileID: 4287259760232714484}
   - component: {fileID: 4287259760229290196}
   - component: {fileID: 4287259760230289940}
-  - component: {fileID: 4287259760239689492}
-  m_Layer: 0
+  - component: {fileID: 61514093683126116}
+  m_Layer: 8
   m_Name: platformFloating (2)
   m_TagString: Untagged
   m_Icon: {fileID: 0}
@@ -315,25 +315,32 @@ MeshRenderer:
   m_SortingLayerID: 0
   m_SortingLayer: 0
   m_SortingOrder: 0
---- !u!95 &4287259760239689492
-Animator:
-  serializedVersion: 3
+--- !u!61 &61514093683126116
+BoxCollider2D:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 4287259760232287444}
   m_Enabled: 1
-  m_Avatar: {fileID: 9000000, guid: 664f8c3657d7bda46b0c0bd1a704f990, type: 3}
-  m_Controller: {fileID: 0}
-  m_CullingMode: 0
-  m_UpdateMode: 0
-  m_ApplyRootMotion: 0
-  m_LinearVelocityBlending: 0
-  m_WarningMessage: 
-  m_HasTransformHierarchy: 1
-  m_AllowConstantClipSamplingOptimization: 1
-  m_KeepAnimatorControllerStateOnDisable: 0
+  m_Density: 1
+  m_Material: {fileID: 0}
+  m_IsTrigger: 0
+  m_UsedByEffector: 0
+  m_UsedByComposite: 0
+  m_Offset: {x: 0, y: 0}
+  m_SpriteTilingProperty:
+    border: {x: 0, y: 0, z: 0, w: 0}
+    pivot: {x: 0, y: 0}
+    oldSize: {x: 0, y: 0}
+    newSize: {x: 0, y: 0}
+    adaptiveTilingThreshold: 0
+    drawMode: 0
+    adaptiveTiling: 0
+  m_AutoTiling: 0
+  serializedVersion: 2
+  m_Size: {x: 2, y: 0.2}
+  m_EdgeRadius: 0
 --- !u!1 &4287259760232287446
 GameObject:
   m_ObjectHideFlags: 0
