+++ From Nothing/Assets/Scripts/Enemy_01_movement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ef93bd9c0d7cf6647ba7d3814f133ac9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
