+++ From Nothing/Assets/Scripts/Level02Dialogue.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 561988693b3a6ed42a9d6552aa8b9769
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
