+++ From Nothing/Assets/Scripts/LevelStart2.cs
@@ -4,10 +4,19 @@ using UnityEngine;
 
 public class LevelStart2 : MonoBehaviour
 {
+    public GameObject player;
+    public GameObject fade;
+
     // Start is called before the first frame update
-    void Start()
+    IEnumerator Start()
     {
-        GetComponent<Animation>().Play("FadeToClear");
+        fade.GetComponent<Animation>().Play("FadeIn");
+        PlayerController.canJump = true;
+        PlayerController.footprintActive = false;
+        yield return new WaitForSeconds(1);
+        GameObject.Find("door01").GetComponent<Level_01_Interact_Manager>().exitEnablePlayer = false;
+        GameObject.Find("door01").GetComponent<Level_01_Interact_Manager>().DoorEntered();
+        GameObject.Find("door01").GetComponent<Level_01_Interact_Manager>().DoorOpen();
     }
 
     // Update is called once per frame
