--- From Nothing/Assets/Scripts/Level_01.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b5cb0474e908ba64cb8dcc471630a8e1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
