+++ From Nothing/Assets/Scripts/Level_01_Interact_Manager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b0a27dde03f4d4b40be340130d0f3ec7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
