+++ From Nothing/Library/LastSceneManagerSetup.txt
@@ -1,4 +1,4 @@
 sceneSetups:
-- path: Assets/Scenes/scLevel2.unity
+- path: Assets/Scenes/scLevel1.unity
   isLoaded: 1
   isActive: 1
