+++ From Nothing/Library/StateCache/SceneView/303267-mainStage.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":false,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true},"in2DMode":false,"pivot":{"x":-3.091169834136963,"y":0.8884196281433106,"z":-3.235812187194824},"rotation":{"x":-0.21009767055511475,"y":-0.034604523330926898,"z":0.0074411313980817799,"w":-0.9770395159721375},"size":13.419447898864746,"orthographic":false}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":false,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true},"in2DMode":false,"pivot":{"x":-21.958965301513673,"y":0.5718053579330444,"z":-2.116107940673828},"rotation":{"x":0.24169662594795228,"y":0.12794271111488343,"z":-0.03216757997870445,"w":0.9613421559333801},"size":3.3933920860290529,"orthographic":false}
